requirements.txt
django-chartwerk==0.5.16
psycopg2==2.7.3.1
celery==4.1.0
redis==2.10.6
gunicorn==19.7.1